pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "antibiotic-stewardship-env"
version = "0.1.0"
description = "Antibiotic Stewardship Environment"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn==0.30.1",
    "pydantic==2.7.4",
    "openenv-core==0.1.1",
    "openai==1.33.0",
    "requests>=2.31.0"
]
